新增api-controller 與修改 user-controller
