+++ routes/modules/user.js
@@ -4,7 +4,7 @@ const userController = require('../../controllers/user-controller')
 const upload = require('../../middleware/multer')
 const { authenticated } = require('../../middleware/auth')
 
-router.put('/:id/edit', authenticated, upload.fields([{name: 'cover', maxCount: 1}, {name: 'avatar', maxCount: 1}]), userController.putUserProfile)
+router.put('/:id/edit', authenticated, upload.fields([{name: 'cover', maxCount: 1}, {name: 'avatar', maxCount: 1}]), userController.putUser)
 router.get('/:id/tweets', authenticated, userController.getUser)
 router.get('/:id/replies', authenticated, userController.getReplies)
 router.get('/:id/likes', authenticated, userController.getLikes)
